- Huge pile of cleanups for compiler warnings mostly from Philipp Thomas and   Ralf Corsepius - Add missing Size() implementation for repomd index

+++ lua/llimits.h
@@ -69,7 +69,7 @@ typedef unsigned char lu_byte;
 ** this is for hashing only; there is no problem if the integer
 ** cannot hold the whole pointer value
 */
-#define IntPoint(p)  ((lu_hash)(p))
+#define IntPoint(p)  ((lu_hash)(unsigned long)(p))
